+++ src/readyToLandDiffs.ts
@@ -16,12 +16,126 @@ class DiffItem implements vscode.QuickPickItem {
   }
 }
 
+const getKeys = <T extends {}>(o: T): Array<keyof T> =>
+  <Array<keyof T>>Object.keys(o);
+
+const joinCommaAnd = (input: string[]) =>
+  input.reduce(
+    (acc, curr, i) =>
+      i !== 0
+        ? acc + `${input.length - 1 === i ? ` and ` : ", "}` + curr
+        : curr,
+    ""
+  );
+
+const buildStatusToIcon = Object.freeze({
+  passed: `$(check)`,
+  busy: `$(loading)`,
+  failed: `$(error)`,
+  unknown: `$(question)`
+});
+type BuildStatus = keyof typeof buildStatusToIcon;
+
 const diffToUrl: { [diffPHID: string]: string } = {};
-let readyToLandDiffs: {
+
+type DiffList = {
   label: string;
   phid: string;
   uri: string;
+  status: BuildStatus;
 }[];
+let acceptedDiffs: DiffList;
+
+const updateStatusBar = (diffsList: DiffList) => {
+  const counts = diffsList.reduce(
+    (acc, value) => ({
+      ...acc,
+      [value.status]: acc[value.status] ? acc[value.status] + 1 : 1
+    }),
+    <
+      {
+        [K in BuildStatus]: number;
+      }
+    >{}
+  );
+
+  statusBar.text(
+    getKeys(buildStatusToIcon)
+      .reduce(
+        (acc, value) =>
+          counts[value]
+            ? `${acc} ${buildStatusToIcon[value]} ${counts[value]}`
+            : acc,
+        ""
+      )
+      .trim()
+  );
+  statusBar.get().tooltip = `Phabricator: ${joinCommaAnd(
+    getKeys(buildStatusToIcon).reduce(
+      (acc, value) =>
+        counts[value] ? [...acc, `${counts[value]} ${value}`] : acc,
+      <string[]>[]
+    )
+  )} ${diffsList.length > 1 ? "diffs" : "diff"}`;
+};
+
+const buildStatus: (
+  diffPHIDs: string[]
+) => Promise<{
+  [diffPHID: string]: BuildStatus;
+} | null> = async diffPHIDs => {
+  try {
+    const response: {
+      result?: {
+        data?: {
+          fields: {
+            objectPHID: string;
+            buildableStatus: {
+              value: "passed" | "preparing" | "buidling" | "failed";
+            };
+          };
+        }[];
+      };
+    } = await request.get({
+      method: "harbormaster.buildable.search",
+      fields: diffPHIDs.reduce(
+        (accumulator, currentValue) => ({
+          ...accumulator,
+          [`constraints[objectPHIDs][${
+            Object.entries(accumulator).length
+          }]`]: currentValue
+        }),
+        {}
+      )
+    });
+
+    // Combine "building" and "preparing" status into "busy"
+    const convertToBusy = (status: string) =>
+      ["preparing", "building"].includes(status) ? "busy" : status;
+
+    return (
+      response.result?.data?.reduce(
+        (acc, value) => ({
+          ...acc,
+          [value.fields.objectPHID]: convertToBusy(
+            value.fields.buildableStatus.value
+          )
+        }),
+        {}
+      ) || null
+    );
+  } catch (e) {
+    console.error(e);
+    log.append(e.message);
+    track.event({
+      category: "Error",
+      action: "Error",
+      label: "Phabricator: could not fetch build status"
+    });
+
+    return null;
+  }
+};
 
 async function list() {
   track.event({
@@ -31,10 +145,10 @@ async function list() {
   });
 
   const selectedItem = await vscode.window.showQuickPick(
-    readyToLandDiffs.map(
-      ({ label, uri }) =>
+    acceptedDiffs.map(
+      ({ status, label, uri }) =>
         new DiffItem({
-          label,
+          label: `${buildStatusToIcon[status]} ${label}`,
           uri
         })
     ),
@@ -58,7 +172,7 @@ async function update(initialLoad: boolean = false) {
   try {
     const baseUrl = await configuration.baseUrl();
     const storeInstance = store.get({ id: baseUrl });
-    const isFirstUndefinedFetch = typeof readyToLandDiffs === "undefined";
+    const isFirstUndefinedFetch = typeof acceptedDiffs === "undefined";
 
     if (initialLoad) {
       statusBar.text(`$(loading)`);
@@ -100,6 +214,7 @@ async function update(initialLoad: boolean = false) {
     }
 
     const diffPHIDs = acceptedRevisions.map(el => el.fields.diffPHID);
+    const acceptedRevisionsBuildstatus = await buildStatus(diffPHIDs);
     const diffPHIDsWithoutLinks = diffPHIDs.filter(phid => !diffToUrl[phid]);
 
     if (diffPHIDsWithoutLinks && diffPHIDsWithoutLinks.length) {
@@ -122,7 +237,6 @@ async function update(initialLoad: boolean = false) {
 
       if (!diffsInfoResponse || !diffsInfoResponse.result) {
         const errorMessage = "[Phabricator] Could not fetch diff URIs";
-        vscode.window.showErrorMessage(errorMessage);
         log.append(errorMessage);
         return;
       }
@@ -132,21 +246,25 @@ async function update(initialLoad: boolean = false) {
       }
     }
 
-    const diffsList = acceptedRevisions.map(el => ({
+    const diffsList: DiffList = acceptedRevisions.map(el => ({
       phid: el.fields.diffPHID,
       label: el.fields.title,
-      uri: diffToUrl[el.fields.diffPHID]
+      uri: diffToUrl[el.fields.diffPHID],
+      status: acceptedRevisionsBuildstatus?.[el.fields.diffPHID] || "unknown"
     }));
-    const previousAcceptedDiffs = [...(readyToLandDiffs || [])];
-    readyToLandDiffs = diffsList;
+    const previousReadyToLandDiffs = [
+      ...(acceptedDiffs || []).filter(el => el.status === "passed")
+    ];
+    acceptedDiffs = diffsList;
 
-    statusBar.text(readyToLandDiffs.length);
-    statusBar.get().tooltip = `Phabricator: ${readyToLandDiffs.length} diffs ready to land`;
+    updateStatusBar(diffsList);
 
     if (!isFirstUndefinedFetch && (await configuration.diffNotifications())) {
       diffsList
         .filter(
-          value => !previousAcceptedDiffs.find(item => item.phid === value.phid)
+          value =>
+            value.status === "passed" &&
+            !previousReadyToLandDiffs.find(item => item.phid === value.phid)
         )
         .forEach(async diff => {
           const open = "Open Diff";
@@ -177,7 +295,7 @@ async function update(initialLoad: boolean = false) {
     track.event({
       category: "Error",
       action: "Error",
-      label: e.message
+      label: "Phabricator: could not load accepted diffs"
     });
   }
 }